Initial commit.  Functional, incomplete.
This is the first draft of a nearly-complete module for serving additional headers contained in files, put together as proof-of-concept over a few days while learning nginx internals.
Implementation is incomplete: some specific headers are not yet handled, some spots of code are stubbed in and need further research, and there is no formal framework for testing.